requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
# psycopg2-binary==2.9.10  # Problemas con encoding en Windows
psycopg[binary]==3.1.18  # PostgreSQL driver moderno (psycopg3)
# sqlite3 viene incluido en Python

# Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
